svelte(phase 3): absorb live-events/live-scroll/live-stats into LiveReload
Move the three LiveReload-exclusive SSE/scroll/stats primitive modules into
LiveReload.svelte's <script module> (exported so their unit tests drive them
directly); the instance onMount calls them. Drop the redundant isAtBottom/
scrollToBottom/scrollElementAboveComposer onMount wrappers — they'd have recursed
against the absorbed module-scope functions.

chat-preview.js stays a shared module (also used by <BtwPopup>). The three
modules are deleted; their tests move to components/session/ with imports
repointed to ./LiveReload.svelte. The two export_html follow/scroll + SSE source
guards now read LiveReload.svelte.
