+++ docs/dev/svelte-migration-plan.md
@@ -609,6 +609,7 @@ per increment.
 
 | Status | Item |
 |---|---|
+| ✅ DONE + verified | **`live-events`/`live-scroll`/`live-stats` → `<LiveReload>` `<script module>`.** Absorbed the three LiveReload-exclusive SSE/scroll/stats primitive modules (~290 LOC) into `LiveReload.svelte`'s `<script module>` (exported, so their unit tests drive them directly); the instance `onMount` calls them. Removed the now-redundant `isAtBottom`/`scrollToBottom`/`scrollElementAboveComposer` onMount wrappers (they'd have recursed against the absorbed module-scope functions — same hazard as the thinking selector). **`chat-preview.js` deliberately KEPT** as a shared module — it's also imported by `<BtwPopup>`. **3 modules DELETED**; their tests moved to `components/session/` with imports repointed to `./LiveReload.svelte` (all 10 cases green). The two `export_html` follow/scroll + SSE source guards repointed to read `LiveReload.svelte` (which now contains that source). Verified: web 531 + knip clean + build (no a11y) + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (live-reload/chat/btw/context-usage green; full **57/2**). |
 | ✅ DONE + verified | **`session-entry-renderer.js` (638 LOC) → `<SessionEntry>` + `<ToolCall>` + `<ToolOutput>` + `<AskQuestion>`.** The message-pane renderer is now real Svelte components (shared live + export): `<SessionEntry>` dispatches entry types declaratively, `<ToolCall>` dispatches tool names, `<ToolOutput>` is the expandable output, `<AskQuestion>` is the interactive ask-user-question card. `{@html}` is used **only** for markdown (`safeMarkedParse`) + pre-rendered ANSI custom-tool HTML — everything else is escaped template, so **no `.js` does innerHTML view rendering** (criterion 1 met for the message pane). Pure helpers split out to `render/entry-format.js`; the download/share/copy utilities to `render/session-entry-actions.js` (used by the live runtime + export). `<SessionContent>` renders `<SessionEntry>` from the model (no injected `renderEntry`); code highlighting stays a post-render pass (live: `applyLazyHighlighting`; export: a synchronous `highlightPending` in `afterRender`). **`session-entry-renderer.js` + test DELETED** (render tests → `SessionEntry`/`ToolCall` component tests; share/copy → `session-entry-actions.test.js`). Two Go guards rewritten: the toggle-markup guard → `SessionEntry`/`ToolOutput`; the ask-question guards → `ToolCall`/`AskQuestion`. **Bug caught by e2e:** the live-only fork/label buttons were gated by a DOM probe evaluated at entry-creation, but `<SessionEntry>` mounts before `<ChatComposer>` → buttons missing; fixed by passing a `live` prop (SessionPage `true`, export default `false`). Verified: web 531 + knip clean + build (no a11y) + `go test ./internal/ui` (incl. export self-containment guard) + go vet + Desktop-Chrome e2e (session-view/annotations/labels/chat/share green; full **57/2**). Annotation offsets re-anchor correctly into the component-rendered DOM. |
 | ✅ DONE + verified | **Four chat selectors → `<ChatComposer>` `<script module>`.** Absorbed `model-selector.js`, `thinking-selector.js`, `slash-command.js`, `mention-autocomplete.js` (≈808 LOC) into ChatComposer's module script as exported functions. `runChatComposer`'s selector params now **default** to the absorbed setups (`modelSelector = { setupModelSelector }`, etc.) so production uses them while the runner's tests still inject mocks. The pure `chat-selectors.js` helpers stay a module (imported by the absorbed code). **4 modules DELETED**; their 4 test files moved to `components/session/` with the import path repointed to `./ChatComposer.svelte` — all 54 selector cases + 18 runner cases green. **Bug caught by e2e:** the runner's internal thinking wrapper was named `setupThinkingLevelSelector` — identical to the absorbed export — so the param default `{ setupThinkingLevelSelector }` bound to the wrapper → infinite recursion (slash/mention setup never ran). Fixed by renaming the wrapper to `loadThinkingSelector` (matching `loadModelSelector`/`loadSlashSelector`/`loadMentionSelector`). Verified: web 527 + knip clean + build (no a11y) + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (slash/mention/composer-keyboard/chat/context-usage green; full **57/2**). |
 | ✅ DONE + verified | **`chat-composer-runner.js` (1121 LOC) → `<ChatComposer>`.** Moved `runChatComposer` verbatim into `ChatComposer.svelte`'s `<script module>` (exported, keeping its full DI signature); the instance `<script>` `onMount` calls it exactly as before. The model/thinking/slash/mention selectors + `chat-api` stay as injected helper modules (passed in). Module-scope `icon`/`t`/`X`/`TextQuote` feed the runner; the markup reuses them. **`chat-composer-runner.js` DELETED**; its 586-LOC / 18-case behavioural suite moved to `components/session/ChatComposer.test.js` with only the import path repointed to `./ChatComposer.svelte` (the `<script module>` export) — **zero test rewrites**, all 18 still green. Verified: web 527 + knip clean + build (no a11y) + autofixer clean + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (chat/slash/mention/context-usage/composer-keyboard/btw green; full **57/2**). |
