+++ web/src/components/session/live-events.test.js
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from 'vitest';
-import { createSessionEventSource, getSessionIdFromLocation, handleSessionReload, wireSessionEvents } from './live-events.js';
+import { createSessionEventSource, getSessionIdFromLocation, handleSessionReload, wireSessionEvents } from './LiveReload.svelte';
 
 describe('live events', () => {
   it('gets session id and creates event source', () => {
