+++ web/src/components/session/live-scroll.test.js
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest';
 import { JSDOM } from 'jsdom';
-import { createFollowButton, removeFollowButton, scrollToBottom, setFollowButtonText, isAtBottom } from './live-scroll.js';
+import { createFollowButton, removeFollowButton, scrollToBottom, setFollowButtonText, isAtBottom } from './LiveReload.svelte';
 
 describe('live scroll helpers', () => {
   it('creates, labels, and removes follow button', () => {
